+++ packages/ui/RadioDecorator/RadioDecorator.js
@@ -1,6 +1,6 @@
 /**
  * Exports the {@link ui/RadioDecorator.RadioDecorator} and
- * {@link ui/RadioDecorator.RadioControllerDecorator} Higher-order Components (HOCs).
+ * {@link ui/RadioDecorator.RadioControllerDecorator} higher-order components (HOCs).
  *
  * @module ui/RadioDecorator
  */
@@ -47,7 +47,7 @@ const defaultConfig = {
 };
 
 /**
- * {@link ui/RadioDecorator.RadioDecorator} is a Higher-order Component that allows another
+ * {@link ui/RadioDecorator.RadioDecorator} is a higher-order component (HOC) that allows another
  * component to have a mutually exclusive relationship with other descendants of the same
  * {@link ui/RadioDecorator.RadioControllerDecorator}.
  *
@@ -102,8 +102,8 @@ const RadioDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			}
 		}
 
-		/**
-		 * Invoked by a RadioControllerDecorator when the wrapped component should be deactivated
+		/*
+		 * Invoked by a `RadioControllerDecorator` when the wrapped component should be deactivated
 		 *
 		 * @returns {undefined}
 		 */