+++ packages/moonstone/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact moonstone module, newest changes on the top.
 
+## [unreleased]
+
+## Fixed
+
+- `moonstone/Dialog` read order of dialog contents
+
 ## [2.0.0] - 2018-07-30
 
 ### Added
@@ -18,7 +24,6 @@ The following is a curated list of changes in the Enact moonstone module, newest
 - `moonstone/VirtualList.VirtualGridList` and `moonstone/VirtualList.VirtualList` to handle focus properly via page up/down keys when switching to 5-way mode
 - `moonstone/Popup` to spot the content after it's mounted
 - `moonstone/Scroller`, `moonstone/VirtualList.VirtualGridList`, and `moonstone/VirtualList.VirtualList` to scroll properly via voice control in RTL locales
-- `moonstone/Dialog` to read properly with reading order of dialog prop `buttons` and `children`
 
 ## [2.0.0-rc.3] - 2018-07-23