+++ packages/moonstone/CHANGELOG.md
@@ -7,6 +7,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 ### Fixed
 
 - `moonstone/Scroller` to go to next page properly via page up/down keys
+- `moonstone/Tooltip` to transform tooltipArrow by 1px up to avoid gaps between its body and arrow
 
 ## [2.0.0] - 2018-07-30